Refactor ProfileScreen to use ProfileHeader widget for improved code organization and readability.
