+++ lib/features/profile/profile.dart
@@ -1,0 +1,2 @@
+export 'profile_screen.dart';
+export 'profile_header.dart';
